add two files

+++ memberws/src/main/java/au/com/new1step/apps/vip/rs/dao/MemberDao.java
@@ -6,7 +6,7 @@ import au.com.new1step.apps.vip.rs.model.Member;
 
 
 public interface MemberDao {
-	public Member fetchMemberById(Integer memId) throws MemberRsException;
+	public Member fetchMemberById(Long memId) throws MemberRsException;
 	public List<Member> fetchMembersByName(Member member) throws MemberRsException;
 	public List<Member> fetchMembersByPhone(Member member) throws MemberRsException;
 	public List<Member> fetchAllMembers(String appId);
@@ -16,7 +16,7 @@ public interface MemberDao {
 	public List<Member> fetchAllActiveSilverMembers(String appId);
 	public List<Member> fetchAllActiveDiamondMembers(String appId);
 	public List<Member> fetchAllActiveOrdinaryMembers(String appId);
-	public Integer insertMember(Member member) throws MemberRsException;
+	public Long insertMember(Member member) throws MemberRsException;
 	public void updateMember(Member member) throws MemberRsException;
 	public void deleteMember(Member member) throws MemberRsException;
 }
