+++ memberws/src/main/java/au/com/new1step/apps/vip/rs/service/impl/MemberServiceImpl.java
@@ -192,7 +192,7 @@ public class MemberServiceImpl implements MemberService {
 	public MemberResponse insertMember(MemberRequest request){
 		MemberResponse response = new MemberResponse();
 		try{
-		     int id = memberDao.insertMember(request.getMember());
+		     Long id = memberDao.insertMember(request.getMember());
 		     List<Member> members = new ArrayList<Member>();
 		     Member member = new Member();
 		     member.setMemId(id);
